requirements.txt
Django>=3.2.6,<3.3.0
djangorestframework>=3.12.4,<3.13.0

pandas>=1.3.3,<1.4.0

gunicorn>=20.1.0,<20.2.0

flake8>=3.9.2,<3.10.0
